Changed bufferBytes to size_t

+++ platforms/opencl/src/OpenCLContext.cpp
@@ -532,9 +532,9 @@ void OpenCLContext::initialize() {
             energyParamDerivBuffer.initialize<cl_float>(*this, numEnergyParamDerivs*energyBufferSize, "energyParamDerivBuffer");
         addAutoclearBuffer(energyParamDerivBuffer);
     }
-    int bufferBytes = max(max((int) velm.getSize()*velm.getElementSize(),
-            energyBufferSize*energyBuffer.getElementSize()),
-            (int) longForceBuffer.getSize()*longForceBuffer.getElementSize());
+    size_t bufferBytes = max(max(velm.getSize()*velm.getElementSize(),
+            (size_t)energyBufferSize*energyBuffer.getElementSize()),
+            longForceBuffer.getSize()*longForceBuffer.getElementSize());
     pinnedBuffer = new cl::Buffer(context, CL_MEM_ALLOC_HOST_PTR, bufferBytes);
     pinnedMemory = currentQueue.enqueueMapBuffer(*pinnedBuffer, CL_TRUE, CL_MAP_READ | CL_MAP_WRITE, 0, bufferBytes);
     for (int i = 0; i < numAtoms; i++) {
